Updated PORT guidance
Clarified the documentation to reflect that the PORT environment variable will always be necessary (during preview) since under the hood, applications will run in a (default) container which also gets a random PORT number assigned. It's also not the container that needs to be modified but the application running within that container.